refactor: consolidate date/time formatting (4 files → 1)
Merge date-formatting.ts, date-utils.ts, time-formatting.ts, and
time-formatters.ts into a single date-time.ts. Flatten dateUtils
object to named exports. Update ~23 consumer imports.

Phase 1B of codebase redundancy cleanup.
