+++ components/forecast/swell-event-card.tsx
@@ -13,7 +13,7 @@ import { useEffect, useState, useRef } from "react";
 import { cn } from "@/lib/utils";
 import { Waves } from "lucide-react";
 import type { SwellEvent } from "@/lib/utils/regional-forecast-utils";
-import { formatCompactDate, formatShortDate } from "@/lib/utils/time-formatters";
+import { formatCompactDate, formatShortDate } from "@/lib/utils/date-time";
 import { getWaveSizeLabel } from "@/lib/utils/wave-formatters";
 import { formatWaveHeightRangeString } from "@/lib/utils/wave-height-formatter";
 import { SwellWaveChart } from "./swell-wave-chart";
